# bpo-30156: Remove property_descr_get() optimization