@vstinner 2018-09-24
I rewrote my commit to add a NEWS entry and update the commit message.