@vstinner 2018-10-01
Thanks for the review @rhettinger ;-)